chore: apply review feedback  - move all function of `wasmplus/keeper/keeper_extension.go` to `wasmplus/keeper/keeper.go`  - remove no need files (keeper_extension.go, metrics.go)
